changed totalNotCity from int to long

+++ src/main/java/com/napier/sem/ReportUtil.java
@@ -24,7 +24,7 @@ public class ReportUtil {
                 pop.reportName = resultSet.getString("reportName");
                 pop.totalCity = resultSet.getInt("totalCity");
                 pop.totalPopulation = resultSet.getLong("totalPopulation");
-                pop.totalNotCity = resultSet.getInt("totalNotCity");
+                pop.totalNotCity = resultSet.getLong("totalNotCity");
                 pop.percentageCity = roundedPercentage(pop.totalCity, pop.totalPopulation);
                 pop.percentageNotCity = roundedPercentage(pop.totalNotCity, pop.totalPopulation);
                 result.add(pop);
@@ -37,35 +37,6 @@ public class ReportUtil {
         }
         return result;
     }
-    // /**
-    //  * Generates a report for the overall population as well as those that live in a city and don't within a continent.
-    //  * @param connection the connection to the database
-    //  * @return
-    //  */
-    // public static ArrayList<Population> peopleDistributionContinent(Connection connection){
-    //     ArrayList<Population> result = new ArrayList<Population>();
-    //     try {
-    //         String[] params = {};
-    //         ResultSet resultSet = SQLUtil.run(connection, "peopleDistribution.sql", params);
-    //         while (resultSet.next()) {
-    //             Population pop = new Population();
-    //             pop.reportName = resultSet.getString("reportName");
-    //             pop.totalCity = resultSet.getInt("totalCity");
-    //             BigDecimal decTotalPop = resultSet.getBigDecimal("totalPopulation");
-    //             pop.totalPopulation = (decTotalPop == null ? null : decTotalPop.toBigInteger());
-    //             pop.totalNotCity = resultSet.getInt("totalNotCity");
-    //             pop.percentageCity = roundedPercentage(pop.totalCity, pop.totalPopulation.intValue());
-    //             pop.percentageNotCity = roundedPercentage(pop.totalNotCity, pop.totalPopulation.intValue());
-    //             result.add(pop);
-    //         }
-    //         resultSet.close();
-    //     }
-    //     catch(Exception e) {
-    //         System.out.println(e.getMessage());
-    //         System.out.println("Failed to get population details");
-    //     }
-    //     return result;
-    // }
 
     public static ArrayList<Population> peopleDistributionContinent(Connection connection){
         ArrayList<Population> result = new ArrayList<Population>();
@@ -77,7 +48,7 @@ public class ReportUtil {
                 pop.reportName = resultSet.getString("reportName");
                 pop.totalCity = resultSet.getInt("totalCity");
                 pop.totalPopulation = resultSet.getLong("totalPopulation");
-                pop.totalNotCity = resultSet.getInt("totalNotCity");
+                pop.totalNotCity = resultSet.getLong("totalNotCity");
                 pop.percentageCity = roundedPercentage(pop.totalCity, pop.totalPopulation);
                 pop.percentageNotCity = roundedPercentage(pop.totalNotCity, pop.totalPopulation);
                 result.add(pop);
